# docs/api-spec.yml
openapi: 3.0.3
info:
  title: Directorio Concón API
  description: |
    API REST para el Directorio de Lugares de Concón (NestJS + Firebase Firestore).
    Contrato entre frontend y backend — actualizar ANTES de implementar cambios.
  version: "1.0.0"
  contact:
    name: Agencia Digital
    url: https://agencia-digital.cl
    email: desarrollo@agencia-digital.cl
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
  - url: https://api.directorio-concon.com/v1
    description: Producción (VPS Docker)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de Firebase Auth (idToken)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Rol:
      type: string
      enum: [admin, owner, member]
      description: |
        Closed `Rol` enum, defined at
        `backend/src/modules/usuarios/domain/rol.enum.ts`. The legacy values
        `'admin' | 'empresa' | 'usuario'` (roles-rename) were renamed to
        `'admin' | 'owner' | 'member'`. Default `'member'` on registration.
...
    Usuario:
      type: object
      description: Document in the Firestore `usuarios` collection.
      properties:
        id:
          type: string
          description: Firebase Auth UID (PK).
        email:
          type: string
          format: email
          description: UNIQUE email. Validated server-side on create.
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          description: Display name.
        rol:
          $ref: '#/components/schemas/Rol'
        placeId:
          type: string
          nullable: true
          description: |
            Reference to the `places` document the user owns. Set when and
            only when `rol === 'owner'`.
        telefono:
          type: string
          nullable: true
          description: Chilean-format phone (free string).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUsuario:
      type: object
      description: |
        Admin-only request to provision a `usuarios` document for an existing
        Firebase Auth user. Self-registration is deferred (see change
        `usuarios-self-signup`).
      required: [email, nombre]
      properties:
        email:
          type: string
          format: email
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        rol:
          $ref: '#/components/schemas/Rol'
          default: member
        placeId:
          type: string
          nullable: true
        telefono:
          type: string
          nullable: true
    UpdatePerfil:
      type: object
      description: |
        Self-service update for the authenticated user's profile. `rol` and
        `placeId` are NOT accepted — they are admin-only (`UpdateRol` /
        admin provisioning). The global `ValidationPipe` with
        `forbidNonWhitelisted: true` rejects any other field with `400`.
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        telefono:
          type: string
          nullable: true
    UpdateRol:
      type: object
      description: |
        Admin-only mutation of a user's `rol`. Validated against the closed
        `Rol` enum server-side; any value outside `[admin, owner, member]`
        receives `400` with message `rol must be one of: admin, owner, member`.
      required: [rol]
      properties:
        rol:
          $ref: '#/components/schemas/Rol'
...
security:
  - bearerAuth: []
paths:
...
  /usuarios/me:
    get:
      summary: Get the authenticated user's own profile
      security:
        - bearerAuth: []
      description: |
        Returns the `usuarios` document matching the verified JWT `uid`. Any
        authenticated role (`'admin'`, `'owner'`, `'member'`) can access
        their own profile — no `@Roles` enforcement.
      responses:
        "200":
          description: The caller's `Usuario` document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: |
            The verified Firebase Auth UID has no matching `usuarios`
            document (orphan — needs admin provisioning before accessing
            any authenticated endpoint).
          $ref: '#/components/schemas/Error'
    put:
      summary: Update the authenticated user's own profile (self-service)
      security:
        - bearerAuth: []
      description: |
        Self-service update of `nombre` and `telefono`. The body accepts
        ONLY `{ nombre?, telefono? }`. `rol` and `placeId` are NOT accepted
        — they are admin-only. The global `ValidationPipe` with
        `forbidNonWhitelisted: true` rejects any other field with `400`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePerfil'
      responses:
        "200":
          description: Updated `Usuario` document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        "400":
          $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: Orphan user (no `usuarios` document).
          $ref: '#/components/schemas/Error'
  /usuarios:
    get:
      summary: List all usuarios (admin-only)
      security:
        - bearerAuth: []
      description: |
        Admin-only. Returns all `usuarios` documents. `@Roles('admin')`
        enforces the role; non-admins receive `403`.
      parameters:
        - name: rol
          in: query
          schema: { $ref: '#/components/schemas/Rol' }
          description: Optional filter by rol.
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Usuario'
                  meta:
                    $ref: '#/components/schemas/Pagination'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: Caller's `rol` is not `'admin'`.
          $ref: '#/components/schemas/Error'
    post:
      summary: Provision a new usuarios document (admin-only)
      security:
        - bearerAuth: []
      description: |
        Admin-only. Provisions a `usuarios` document for an existing
        Firebase Auth user (the Firebase Auth signup happens client-side;
        this endpoint writes the matching `usuarios` row). Default
        `rol: 'member'` if omitted. Email uniqueness is enforced server-side
        (`ConflictException`/`409`). Self-registration is deferred to a
        future `usuarios-self-signup` change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUsuario'
      responses:
        "201":
          description: Usuario created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        "400":
          $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: Caller's `rol` is not `'admin'`.
          $ref: '#/components/schemas/Error'
        "409":
          description: A `usuarios` document with that `email` already exists.
          $ref: '#/components/schemas/Error'
  /usuarios/{uid}:
    get:
      summary: Get a usuarios document by UID (admin-only)
      security:
        - bearerAuth: []
      description: Admin-only. `@Roles('admin')` enforces; non-admins receive `403`.
      parameters:
        - name: uid
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuario document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: Caller's `rol` is not `'admin'`.
          $ref: '#/components/schemas/Error'
        "404":
          $ref: '#/components/schemas/Error'
  /usuarios/{uid}/rol:
    put:
      summary: Update a user's rol (admin-only, validates closed enum)
      security:
        - bearerAuth: []
      description: |
        Admin-only. Body MUST be `{ rol: 'admin'|'owner'|'member' }`.
        Validated against `ROL_VALUES` `@IsEnum` — any other value
        receives `400` with message `rol must be one of: admin, owner, member`.
        Non-admins receive `403`.
      parameters:
        - name: uid
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRol'
      responses:
        "200":
          description: Updated `Usuario` document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        "400":
          description: Body's `rol` is not in `[admin, owner, member]`.
          $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/schemas/Error'
        "403":
          description: Caller's `rol` is not `'admin'`.
          $ref: '#/components/schemas/Error'
        "404":
          $ref: '#/components/schemas/Error'

  # — solicitudes approve/reject paths (change auth-usuarios) —
...